mi_proyecto_django/requirements.txt
asgiref==3.10.0
Django==5.2.8
djangorestframework==3.16.1
Faker==37.12.0
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.3
tzdata==2025.2
djangorestframework
python-dotenv
django-allauth
